CRUD implementation

+++ app/Http/Controllers/Api/V1/CompanyController.php
@@ -9,6 +9,9 @@ use Illuminate\Http\Request;
 use App\Http\Resources\Api\V1\CompanyResource;
 use App\Http\Resources\Api\V1\CompanyCollection;
 
+use App\Http\Requests\Company\StoreCompanyRequest;
+use App\Http\Requests\Company\UpdateCompanyRequest;
+
 class CompanyController extends Controller
 {
     /**
@@ -19,7 +22,6 @@ class CompanyController extends Controller
     public function index()
     {
         return new CompanyCollection(Company::all());
-
     }
 
     /**
@@ -28,10 +30,10 @@ class CompanyController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreCompanyRequest $request)
     {
         Company::create($request->all());
-        return response()->json(['status' => true, 'message' => "Pomyślnie dodano dane firmy."], 200);
+        return response()->json(['status' => true, 'message' => "Company data added successfully."], 201);
     }
 
     /**
@@ -53,10 +55,10 @@ class CompanyController extends Controller
      * @param  \App\Models\Company  $company
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Company $company)
+    public function update(UpdateCompanyRequest $request, Company $company)
     {
         $company->update($request->all());
-        return response()->json(['status' => true, 'message' => "Pomyślnie zaktualizowano dane firmy."], 200);
+        return response()->json(['status' => true, 'message' => "Company data updated successfully."], 200);
     }
 
     /**
@@ -65,9 +67,31 @@ class CompanyController extends Controller
      * @param  \App\Models\Company  $company
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Company $company)
+    public function destroy($companyId)
     {
+        $company = Company::find($companyId);
+        if(empty($company)) {
+            return response()->json(['status' => false, 'message' => sprintf("Company with ID: %s not found.", $companyId)], 400);
+        }
+
         $company->delete();
-        return response()->json(['status' => true, 'message' => "Pomyślnie usunięto dane firmy."], 200);
+        return response()->json(['status' => true, 'message' => "Company data deleted successfully."], 200);
+
+    }
+
+    /**
+     * Get employees that belongs to the specified company
+     *
+     * @param  mixed $companyId
+     * @return Object
+     */
+    public function getEmployees($companyId)
+    {
+        $company = Company::find($companyId);
+        if(empty($company)) {
+            return response()->json(['status' => false, 'message' => sprintf("Nie znaleziono firmy o ID: %s", $companyId)], 400);
+        }
+
+        return $company->employees()->get();
     }
 }
