+++ app/Http/Controllers/Api/V1/EmployeeController.php
@@ -9,6 +9,9 @@ use App\Models\Employee;
 use App\Http\Resources\Api\V1\EmployeeResource;
 use App\Http\Resources\Api\V1\EmployeeCollection;
 
+use App\Http\Requests\Employee\StoreEmployeeRequest;
+use App\Http\Requests\Employee\UpdateEmployeeRequest;
+
 class EmployeeController extends Controller
 {
         /**
@@ -28,10 +31,10 @@ class EmployeeController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
+    public function store(StoreEmployeeRequest $request)
     {
         Employee::create($request->all());
-        return response()->json(['status' => true, 'message' => "Pomyślnie dodano dane pracownika."], 200);
+        return response()->json(['status' => true, 'message' => "Employee data added successfully."], 201);
     }
 
     /**
@@ -52,10 +55,10 @@ class EmployeeController extends Controller
      * @param  \App\Models\Employee  $employee
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Employee $employee)
+    public function update(UpdateEmployeeRequest $request, Employee $employee)
     {
         $employee->update($request->all());
-        return response()->json(['status' => true, 'message' => "Pomyślnie zaktualizowano dane pracownika."], 200);
+        return response()->json(['status' => true, 'message' => "Employee data updated successfully."], 200);
     }
 
     /**
@@ -64,9 +67,30 @@ class EmployeeController extends Controller
      * @param  \App\Models\Employee  $employee
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Employee $employee)
+    public function destroy($employeeId)
     {
+        $employee = Employee::find($employeeId);
+        if(empty($employee)) {
+            return response()->json(['status' => false, 'message' => sprintf("Employee with ID: %s not found.", $employeeId)], 400);
+        }
+
         $employee->delete();
-        return response()->json(['status' => true, 'message' => "Pomyślnie usunięto dane pracownika."], 200);
+        return response()->json(['status' => true, 'message' => "Employee data deleted successfully."], 200);
+    }
+
+    /**
+     * Get company that belongs to the specified employee
+     *
+     * @param  mixed $employeeId
+     * @return Object
+     */
+    public function getCompany($employeeId)
+    {
+        $employee = Employee::find($employeeId);
+        if(empty($employee)) {
+            return response()->json(['status' => false, 'message' => sprintf("Nie znaleziono pracownika o ID: %s", $employeeId)], 400);
+        }
+
+        return $employee->company()->first();
     }
 }
